pyproject.toml
[project]
name = "rankforge"
version = "1.0.0"
description = "AI-powered SEO CLI toolkit — keyword research, backlinks, audits, and automation."
requires-python = ">=3.10"
authors = [{ name = "RankForge" }]
license = { text = "MIT" }
readme = "README.md"

dependencies = [
    "typer[all]>=0.12.0",
    "rich>=13.7.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.1.0",
    "openai>=1.30.0",
    "anthropic>=0.34.0",
    "google-generativeai>=0.7.0",
]

[project.optional-dependencies]
vector = ["chromadb>=0.4.0"]

[project.scripts]
rankforge = "rankforge.main:app"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["rankforge*"]
